+++ controllers/ingressClassCreator_test.go
@@ -1,32 +1,19 @@
-package controllers_test
+package controllers
 
 import (
+	"context"
 	"testing"
-
-	"k8s.io/apimachinery/pkg/runtime"
-
-	"github.com/cloudogu/k8s-service-discovery/controllers"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 
 	networking "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	ctrl "sigs.k8s.io/controller-runtime"
 
 	"github.com/stretchr/testify/require"
 	testclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
-func getScheme() *runtime.Scheme {
-	scheme := runtime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	return scheme
-}
-
 func TestNewIngressClassCreator(t *testing.T) {
 	clientMock := testclient.NewClientBuilder().WithScheme(getScheme()).Build()
-	creator := controllers.NewIngressClassCreator(clientMock, "my-ingress-class")
+	creator := NewIngressClassCreator(clientMock, "my-ingress-class")
 
 	require.NotNil(t, creator)
 }
@@ -37,10 +24,10 @@ func TestIngressClassCreator_CreateIngressClass(t *testing.T) {
 
 		// given
 		clientMock := testclient.NewClientBuilder().WithScheme(getScheme()).Build()
-		creator := controllers.NewIngressClassCreator(clientMock, "my-ingress-class")
+		creator := NewIngressClassCreator(clientMock, "my-ingress-class")
 
 		// when
-		err := creator.CreateIngressClass(ctrl.Log.WithName("test"))
+		err := creator.CreateIngressClass(context.Background())
 
 		// then
 		require.NoError(t, err)
@@ -60,10 +47,10 @@ func TestIngressClassCreator_CreateIngressClass(t *testing.T) {
 			},
 		}
 		clientMock := testclient.NewClientBuilder().WithScheme(getScheme()).WithObjects(ingressClass).Build()
-		creator := controllers.NewIngressClassCreator(clientMock, "my-ingress-class")
+		creator := NewIngressClassCreator(clientMock, "my-ingress-class")
 
 		// when
-		err := creator.CreateIngressClass(ctrl.Log.WithName("test"))
+		err := creator.CreateIngressClass(context.Background())
 
 		// then
 		require.NoError(t, err)