+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#10] Support for maintenance mode. See [maintenance mode](docs/operations/maintenance_mode_en.md) for more details.
+
+### Changed
+- [#10] Update `ces-build-lib` to version `1.56.0`
+- [#10] Update `makefiles` to version `7.0.1`
 
 ## [v0.4.0] - 2022-08-29
 ### Added
@@ -17,8 +23,8 @@ A watch recognizes changes in the registry for the certificate and updates the s
 
 ## [v0.2.0] - 2022-06-08
 ### Added
-- Warp menu generation
-  - [#3] Add runnable to the controller which observes keys in the etcd specified in a configmap `k8s-ces-warp-config`
+- [#3] Warp menu generation
+  - Add runnable to the controller which observes keys in the etcd specified in a configmap `k8s-ces-warp-config`
   and creates warp menu entries in `k8s-ces-menu-json` for the nginx-ingress dogu.
 
 ### Changed